test: move bin/multi-path
