+++ test/integration/bin/multi-path/fixtures/.gitignore
