+++ test/integration/bin/multi-path/fixtures/package.json
