+++ test/integration/bin/multi-path/fixtures/src/bin/a.ts
