+++ test/integration/bin/multi-path/fixtures/src/bin/b.ts
